+++ packages/sdk/src/definition/graphql-introspection.ts
@@ -20,6 +20,7 @@ import {
 	applyNamespaceToDirectiveConfiguration,
 	applyNameSpaceToFieldConfigurations,
 	applyNameSpaceToGraphQLSchema,
+	applyNameSpaceToSingleTypeFields,
 	applyNameSpaceToTypeFields,
 	generateTypeConfigurationsForNamespace,
 } from './namespacing';
@@ -80,8 +81,8 @@ export const introspectGraphql = async (introspection: GraphQLIntrospection): Pr
 		let schema = await introspectGraphQLSchema(introspection, introspectionHeaders);
 		schema = lexicographicSortSchema(schema);
 		const federationEnabled = isFederationService(schema);
-		let schemaSDL = cleanupSchema(schema, introspection);
-		schemaSDL = transformSchema.replaceCustomScalars(schemaSDL, introspection);
+		const upstreamSchema = cleanupSchema(schema, introspection);
+		const { schemaSDL, customScalarTypeFields } = transformSchema.replaceCustomScalars(upstreamSchema, introspection);
 		const serviceSDL = !federationEnabled
 			? undefined
 			: introspection.loadSchemaFromString
@@ -142,10 +143,15 @@ export const introspectGraphql = async (introspection: GraphQLIntrospection): Pr
 							Enabled: federationEnabled,
 							ServiceSDL: serviceSDL || '',
 						},
-						UpstreamSchema: schemaSDL,
+						UpstreamSchema: upstreamSchema,
 						HooksConfiguration: {
 							onWSTransportConnectionInit: false,
 						},
+						CustomScalarTypeFields: applyNameSpaceToSingleTypeFields(
+							customScalarTypeFields,
+							graphQLSchema,
+							introspection.apiNamespace
+						),
 					},
 					Directives: applyNamespaceToDirectiveConfiguration(schema, introspection.apiNamespace),
 					RequestTimeoutSeconds: introspection.requestTimeoutSeconds ?? 0,