+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.14](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.3.12...@wundergraph/golang-client@0.3.14) (2022-11-20)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.3.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.3.12...@wundergraph/golang-client@0.3.13) (2022-11-20)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.3.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.3.11...@wundergraph/golang-client@0.3.12) (2022-11-18)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.3.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.3.10...@wundergraph/golang-client@0.3.11) (2022-11-17)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.3.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.3.9...@wundergraph/golang-client@0.3.10) (2022-11-11)
 
 **Note:** Version bump only for package @wundergraph/golang-client